--- src/components/login/RedirectKakao.jsx
@@ -1,14 +1,0 @@
-import { useEffect } from 'react';
-import useKakoAuth from '../../hooks/useKakaoAuth';
-import LoadingSpinner from '../LoadingSpinner';
-
-const RedirectLogin = () => {
-  const { handleLogin } = useKakoAuth();
-  useEffect(() => {
-    handleLogin();
-  }, []);
-
-  return <LoadingSpinner />;
-};
-
-export default RedirectLogin;
